Project 5: Camera Characterization
In this project you will characterize your digital camera so that it can be used to estimate the XYZ values of imaged surfaces.  First, using the image of the ColorChecker chart you created in Project 1, you will extract the RGB values of the color patches, and use these values together with measured XYZ values for the patches to plot your camera’s tone transfer functions (TTFs). You will then fit functions to the TTFs and use them to linearize the camera’s response with respect to relative luminance. You will plot these linearized response functions and create images that visualize the camera’s original and linearized RGB responses to the patches in the ColorChecker chart. You will then use the linearized RGB data for the ColorChecker patches and measured XYZ values for the patches to derive a matrix that transforms linearized camera RGBs to estimated XYZs.  You will then check the accuracy of this transformation by calculating ΔE_{ab} color differences between the ColorMunki-measured  and camera-estimated patch values and visualizing images of the ColorChecker from both sets of values. Finally you will use your camera model to estimate the XYZ and Lab values of your colored patches from the image you created in Project 1 and will compare the ColorMunki-measured and camera-estimated values both numerically and visually.